GUI bet後、サイドバー更新

+++ src/main/java/red/man10/VaultManager.java
@@ -51,7 +51,7 @@ public class VaultManager {
     public void showBalance(UUID uuid){
         OfflinePlayer p = Bukkit.getOfflinePlayer(uuid).getPlayer();
         double money = getBalance(uuid);
-        p.getPlayer().sendMessage(ChatColor.YELLOW + "あなたの所持金は$" + money);
+        p.getPlayer().sendMessage(ChatColor.YELLOW + "あなたの所持金は$" + (int)money);
     }
     /////////////////////////////////////
     //      引き出し
@@ -60,7 +60,7 @@ public class VaultManager {
         OfflinePlayer p = Bukkit.getOfflinePlayer(uuid).getPlayer();
         EconomyResponse resp = economy.withdrawPlayer(p,money);
         if(resp.transactionSuccess()){
-            p.getPlayer().sendMessage(ChatColor.YELLOW + "$"+money+"支払いました");
+            p.getPlayer().sendMessage(ChatColor.YELLOW + "$"+(int)money+"支払いました");
             return true;
         }
         return  false;
@@ -72,7 +72,7 @@ public class VaultManager {
         OfflinePlayer p = Bukkit.getOfflinePlayer(uuid).getPlayer();
         EconomyResponse resp = economy.depositPlayer(p,money);
         if(resp.transactionSuccess()){
-            p.getPlayer().sendMessage(ChatColor.YELLOW + "$"+money+"受取りました");
+            p.getPlayer().sendMessage(ChatColor.YELLOW + "$"+(int)money+"受取りました");
             return true;
         }
         return  false;
